refact: using blueprints to organize the routes
